+++ Get-MarkdownHelp.ps1
@@ -15,7 +15,7 @@ function Get-MarkdownHelp {
     #>
     param(
     # The name of the specified command or concept.
-    [Parameter(Position=0, ValueFromPipelineByPropertyName=$true)]
+    [Parameter(Position=0, ValueFromPipelineByPropertyName)]
     [ValidateNotNullOrEmpty()]
     [string]
     $Name,
@@ -27,17 +27,24 @@ function Get-MarkdownHelp {
     # If set, will interlink documentation as if it were for GitHub pages, beneath a given directory
     [Alias('GitHubPageRoot')]    
     [string]
-    $GitHubDocRoot
-    )
+    $GitHubDocRoot,
 
+    # If provided, will rename the help topic before getting markdown.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string]
+    $Rename,
+
+    # The order of the sections.  If not provided, this will be the order they are defined in the formatter.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string[]]
+    $SectionOrder
+    )
 
     process
     {
-        $paramCopy = @{} + $PSBoundParameters
-        $myParams  = @{} + $PSBoundParameters
-        $paramCopy.Remove('Wiki')
-        $paramCopy.Remove('GitHubDocRoot')
-        $gotHelp = Get-Help @paramCopy 
+        $getHelp = @{name=$Name}
+        $myParams= @{} + $PSBoundParameters        
+        $gotHelp = Get-Help @getHelp 
         if (-not $gotHelp) {
             Write-Error "Could not get help for $name"
             return
@@ -48,14 +55,20 @@ function Get-MarkdownHelp {
                     if ($in -is [string]) {
                         $in
                     } else {
-                        $helpObj = $_                        
+                        $helpObj = $_
+                        if ($Rename) {
+                            $helpObj | Add-Member NoteProperty Rename $Rename -Force
+                        }
                         $helpObj.pstypenames.clear()
                         $helpObj.pstypenames.add('PowerShell.Markdown.Help')
+                        if ($SectionOrder) {
+                            $helpObj | Add-Member NoteProperty SectionOrder $SectionOrder -Force    
+                        }
                         $helpObj | Add-Member NoteProperty WikiLink ($Wiki -as [bool]) -Force
                         if ($myParams.ContainsKey("GitHubDocRoot")) {
                             $helpObj | Add-Member NoteProperty DocLink $GitHubDocRoot -Force
                         }
-                        $helpObj | Out-String -Width 1mb
+                        $helpObj
                     }
                 } 
             }