+++ Save-MarkdownHelp.ps1
@@ -44,7 +44,16 @@ function Save-MarkdownHelp
     # If provided, will replace parts of the names of the scripts discovered in a -ScriptDirectory beneath a module with a given Regex replacement.
     [Parameter(ValueFromPipelineByPropertyName)]
     [string[]]
-    $ReplaceScriptNameWith
+    $ReplaceScriptNameWith,
+
+    # If set, will output changed or created files.
+    [switch]
+    $PassThru,
+
+    # The order of the sections.  If not provided, this will be the order they are defined in the formatter.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string[]]
+    $SectionOrder
     )
 
     begin {
@@ -58,7 +67,10 @@ function Save-MarkdownHelp
     }
 
     process {
-
+        $getMarkdownHelpSplatBase = @{}
+        if ($SectionOrder) {
+            $getMarkdownHelpSplatBase.SectionOrder =$SectionOrder
+        }
         #region Save the Markdowns
         foreach ($m in $Module) { # Walk thru the list of module names.            
             if ($t -gt 1) {
@@ -94,10 +106,13 @@ function Save-MarkdownHelp
 
             foreach ($cmd in $theModule.ExportedCommands.Values) {
                 $docOutputPath = Join-Path $outputPath ($cmd.Name + '.md')
-                $getMarkdownHelpSplat = @{Name="$cmd"}
+                $getMarkdownHelpSplat = @{Name="$cmd"} + $getMarkdownHelpSplatBase
                 if ($Wiki) { $getMarkdownHelpSplat.Wiki = $Wiki}
                 else { $getMarkdownHelpSplat.GitHubDocRoot = "$($outputPath|Split-Path -Leaf)"}
-                & $GetMarkdownHelp @getMarkdownHelpSplat| Set-Content -Path $docOutputPath -Encoding utf8
+                & $GetMarkdownHelp @getMarkdownHelpSplat| Out-String -Width 1mb | Set-Content -Path $docOutputPath -Encoding utf8
+                if ($PassThru) {
+                    Get-Item -Path $docOutputPath -ErrorAction SilentlyContinue
+                }
             }
 
             if ($ScriptPath) {
@@ -109,21 +124,27 @@ function Save-MarkdownHelp
                         Where-Object Extension -eq '.ps1' |
                         ForEach-Object {
                             $ps1File = $_
-                            $getMarkdownHelpSplat = @{Name="$($ps1File.FullName)"}
-
-                            $replacedName = $ps1File.Name
+                            $getMarkdownHelpSplat = @{Name="$($ps1File.FullName)"} + $getMarkdownHelpSplatBase
+                            $replacedFileName = $ps1File.Name
                             @(for ($ri = 0; $ri -lt $ReplaceScriptName.Length; $ri++) {
                                 if ($ReplaceScriptNameWith[$ri]) {
-                                    $replacedName = $replacedName -replace $ReplaceScriptName[$ri], $ReplaceScriptNameWith[$ri]
+                                    $replacedFileName = $replacedFileName -replace $ReplaceScriptName[$ri], $ReplaceScriptNameWith[$ri]
                                 } else {
-                                    $replacedName = $replacedName -replace $ReplaceScriptName[$ri]
+                                    $replacedFileName = $replacedFileName -replace $ReplaceScriptName[$ri]
                                 }
                             })
-                            $docOutputPath = Join-Path $outputPath ($replacedName + '.md')
+                            $docOutputPath = Join-Path $outputPath ($replacedFileName + '.md')
+                            $relativePath = $ps1File.FullName.Substring("$theModuleRoot".Length).TrimStart('/\').Replace('\','/')
+                            $getMarkdownHelpSplat.Rename = $relativePath
                             if ($Wiki) { $getMarkdownHelpSplat.Wiki = $Wiki}
                             else { $getMarkdownHelpSplat.GitHubDocRoot = "$($outputPath|Split-Path -Leaf)"}
-                            & $GetMarkdownHelp @getMarkdownHelpSplat| Set-Content -Path $docOutputPath -Encoding utf8
+                            & $GetMarkdownHelp @getMarkdownHelpSplat| Out-String -Width 1mb | Set-Content -Path $docOutputPath -Encoding utf8
+                            if ($PassThru) {
+                                Get-Item -Path $docOutputPath -ErrorAction SilentlyContinue
+                            }
                         }
+
+                    
                 }
             }
          }