+++ allcommands.ps1
@@ -418,7 +418,7 @@ function Get-MarkdownHelp {
     #>
     param(
     # The name of the specified command or concept.
-    [Parameter(Position=0, ValueFromPipelineByPropertyName=$true)]
+    [Parameter(Position=0, ValueFromPipelineByPropertyName)]
     [ValidateNotNullOrEmpty()]
     [string]
     $Name,
@@ -430,17 +430,24 @@ function Get-MarkdownHelp {
     # If set, will interlink documentation as if it were for GitHub pages, beneath a given directory
     [Alias('GitHubPageRoot')]    
     [string]
-    $GitHubDocRoot
-    )
+    $GitHubDocRoot,
 
+    # If provided, will rename the help topic before getting markdown.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string]
+    $Rename,
+
+    # The order of the sections.  If not provided, this will be the order they are defined in the formatter.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string[]]
+    $SectionOrder
+    )
 
     process
     {
-        $paramCopy = @{} + $PSBoundParameters
-        $myParams  = @{} + $PSBoundParameters
-        $paramCopy.Remove('Wiki')
-        $paramCopy.Remove('GitHubDocRoot')
-        $gotHelp = Get-Help @paramCopy 
+        $getHelp = @{name=$Name}
+        $myParams= @{} + $PSBoundParameters        
+        $gotHelp = Get-Help @getHelp 
         if (-not $gotHelp) {
             Write-Error "Could not get help for $name"
             return
@@ -451,14 +458,20 @@ function Get-MarkdownHelp {
                     if ($in -is [string]) {
                         $in
                     } else {
-                        $helpObj = $_                        
+                        $helpObj = $_
+                        if ($Rename) {
+                            $helpObj | Add-Member NoteProperty Rename $Rename -Force
+                        }
                         $helpObj.pstypenames.clear()
                         $helpObj.pstypenames.add('PowerShell.Markdown.Help')
+                        if ($SectionOrder) {
+                            $helpObj | Add-Member NoteProperty SectionOrder $SectionOrder -Force    
+                        }
                         $helpObj | Add-Member NoteProperty WikiLink ($Wiki -as [bool]) -Force
                         if ($myParams.ContainsKey("GitHubDocRoot")) {
                             $helpObj | Add-Member NoteProperty DocLink $GitHubDocRoot -Force
                         }
-                        $helpObj | Out-String -Width 1mb
+                        $helpObj
                     }
                 } 
             }        
@@ -1385,7 +1398,16 @@ function Save-MarkdownHelp
     # If provided, will replace parts of the names of the scripts discovered in a -ScriptDirectory beneath a module with a given Regex replacement.
     [Parameter(ValueFromPipelineByPropertyName)]
     [string[]]
-    $ReplaceScriptNameWith
+    $ReplaceScriptNameWith,
+
+    # If set, will output changed or created files.
+    [switch]
+    $PassThru,
+
+    # The order of the sections.  If not provided, this will be the order they are defined in the formatter.
+    [Parameter(ValueFromPipelineByPropertyName)]
+    [string[]]
+    $SectionOrder
     )
 
     begin {
@@ -1399,7 +1421,10 @@ function Save-MarkdownHelp
     }
 
     process {
-
+        $getMarkdownHelpSplatBase = @{}
+        if ($SectionOrder) {
+            $getMarkdownHelpSplatBase.SectionOrder =$SectionOrder
+        }
         #region Save the Markdowns
         foreach ($m in $Module) { # Walk thru the list of module names.            
             if ($t -gt 1) {
@@ -1435,10 +1460,13 @@ function Save-MarkdownHelp
 
             foreach ($cmd in $theModule.ExportedCommands.Values) {
                 $docOutputPath = Join-Path $outputPath ($cmd.Name + '.md')
-                $getMarkdownHelpSplat = @{Name="$cmd"}
+                $getMarkdownHelpSplat = @{Name="$cmd"} + $getMarkdownHelpSplatBase
                 if ($Wiki) { $getMarkdownHelpSplat.Wiki = $Wiki}
                 else { $getMarkdownHelpSplat.GitHubDocRoot = "$($outputPath|Split-Path -Leaf)"}
-                & $GetMarkdownHelp @getMarkdownHelpSplat| Set-Content -Path $docOutputPath -Encoding utf8
+                & $GetMarkdownHelp @getMarkdownHelpSplat| Out-String -Width 1mb | Set-Content -Path $docOutputPath -Encoding utf8
+                if ($PassThru) {
+                    Get-Item -Path $docOutputPath -ErrorAction SilentlyContinue
+                }
             }
 
             if ($ScriptPath) {
@@ -1450,21 +1478,27 @@ function Save-MarkdownHelp
                         Where-Object Extension -eq '.ps1' |
                         ForEach-Object {
                             $ps1File = $_
-                            $getMarkdownHelpSplat = @{Name="$($ps1File.FullName)"}
-
-                            $replacedName = $ps1File.Name
+                            $getMarkdownHelpSplat = @{Name="$($ps1File.FullName)"} + $getMarkdownHelpSplatBase
+                            $replacedFileName = $ps1File.Name
                             @(for ($ri = 0; $ri -lt $ReplaceScriptName.Length; $ri++) {
                                 if ($ReplaceScriptNameWith[$ri]) {
-                                    $replacedName = $replacedName -replace $ReplaceScriptName[$ri], $ReplaceScriptNameWith[$ri]
+                                    $replacedFileName = $replacedFileName -replace $ReplaceScriptName[$ri], $ReplaceScriptNameWith[$ri]
                                 } else {
-                                    $replacedName = $replacedName -replace $ReplaceScriptName[$ri]
+                                    $replacedFileName = $replacedFileName -replace $ReplaceScriptName[$ri]
                                 }
                             })
-                            $docOutputPath = Join-Path $outputPath ($replacedName + '.md')
+                            $docOutputPath = Join-Path $outputPath ($replacedFileName + '.md')
+                            $relativePath = $ps1File.FullName.Substring("$theModuleRoot".Length).TrimStart('/\').Replace('\','/')
+                            $getMarkdownHelpSplat.Rename = $relativePath
                             if ($Wiki) { $getMarkdownHelpSplat.Wiki = $Wiki}
                             else { $getMarkdownHelpSplat.GitHubDocRoot = "$($outputPath|Split-Path -Leaf)"}
-                            & $GetMarkdownHelp @getMarkdownHelpSplat| Set-Content -Path $docOutputPath -Encoding utf8
+                            & $GetMarkdownHelp @getMarkdownHelpSplat| Out-String -Width 1mb | Set-Content -Path $docOutputPath -Encoding utf8
+                            if ($PassThru) {
+                                Get-Item -Path $docOutputPath -ErrorAction SilentlyContinue
+                            }
                         }
+
+                    
                 }
             }
          }